fix(quality): extract context-fit evaluation to keep comboStructure.ts under cap
The model_context_override fix grew open-sse/services/combo/comboStructure.ts
past the 800-line file-size cap. Extract evaluateContextLimit() (the
override-then-catalog context-fit check) into a new leaf,
open-sse/services/combo/contextOverrideGate.ts, so comboStructure.ts only
keeps the two call-site wires. No behavior change.
