+++ open-sse/services/combo/comboStructure.ts
@@ -12,13 +12,13 @@
  */
 
 import { getModelContextLimit } from "../../../src/lib/modelCapabilities";
-import { getModelContextOverride } from "../../../src/lib/db/modelContextOverrides";
 import { getComboModelString, normalizeComboStep } from "../../../src/lib/combos/steps.ts";
 import { estimateTokens } from "../contextManager.ts";
 import { getResolvedModelCapabilities } from "../modelCapabilities.ts";
 import { parseModel } from "../model.ts";
 import { dedupeTargetsByExecutionKey, isRecord } from "./comboData.ts";
 import { getTargetProvider, MAX_COMBO_DEPTH } from "./comboPredicates.ts";
+import { evaluateContextLimit } from "./contextOverrideGate.ts";
 import { hasEstimableContent } from "./knownContextOverflow.ts";
 import {
   normalizeModelEntry,
@@ -486,70 +486,6 @@ function exceedsKnownOutputLimit(
 ): boolean {
   if (requestedOutputTokens <= 0 || maxOutputTokens === null) return false;
   return maxOutputTokens < requestedOutputTokens;
-}
-
-/**
- * Decide whether a target's known context limit accommodates the request.
- *
- * `maxInputTokens` is an **input-only** cap — the requested output reserve is
- * already enforced separately against `maxOutputTokens` (see
- * `exceedsKnownOutputLimit`), so it must NOT be re-counted here. Comparing
- * `maxInputTokens` against `estimatedInputTokens + requestedOutputTokens`
- * double-counted the output reserve and shrank the effective input allowance
- * (#7039).
- *
- * `contextWindow` is the total window, so input + output must both fit.
- *
- * Returns `true` when the known limit accommodates the request, `false` when
- * it is known to be too small, and `null` when no limit metadata is known.
- */
-function evaluateContextLimit(
-  capabilities: { maxInputTokens?: number | null; contextWindow?: number | null },
-  requirements: { estimatedInputTokens: number; requiredContextTokens: number },
-  modelStr?: string
-): boolean | null {
-  // A persisted context override (Feature 5004 — operator-set or auto-discovered
-  // real usable window) wins for server-side routing. The catalog's
-  // `maxInputTokens` can be a deliberately smaller *client-facing* hint (e.g. set
-  // below the true window so coding agents auto-compact — #6191); using it to
-  // filter fallback targets wrongly drops otherwise-capable providers for large
-  // prompts, collapsing the pool to one provider and producing a hard 503 with no
-  // fallback once that provider's quota is exhausted. The override reflects the
-  // real capacity, so it supersedes both catalog limits here. Use the raw override
-  // (getModelContextOverride returns null when none is set) — NOT
-  // getModelContextLimitForModelString, which falls back to contextWindow and would
-  // therefore bypass the maxInputTokens cap for every model, not just overridden ones.
-  if (modelStr) {
-    const parsed = parseModel(modelStr);
-    const override = getModelContextOverride(parsed.provider, parsed.model);
-    if (override != null) {
-      return override >= requirements.requiredContextTokens;
-    }
-  }
-
-  const hasMaxInput = capabilities.maxInputTokens != null;
-  const hasContextWindow = capabilities.contextWindow != null;
-
-  // Neither limit is known — cannot judge.
-  if (!hasMaxInput && !hasContextWindow) return null;
-
-  // The input-only cap must accommodate the estimated input.
-  const inputFits = hasMaxInput
-    ? capabilities.maxInputTokens! >= requirements.estimatedInputTokens
-    : true;
-
-  // The total window must accommodate input + requested output. The output
-  // reserve is enforced separately via `maxOutputTokens`, but when a model
-  // exposes both `maxInputTokens` and `contextWindow` the two must not be
-  // checked in isolation: a request whose input fits `maxInputTokens` but whose
-  // input + output exceeds `contextWindow` must still be rejected (#7039
-  // follow-up — shared-window models where `maxInputTokens` defaults to the
-  // total window size).
-  const totalFits = hasContextWindow
-    ? capabilities.contextWindow! >= requirements.requiredContextTokens
-    : true;
-
-  return inputFits && totalFits;
 }
 
 function hasKnownCompatibleContextLimit(
